Move PolynomialRingElement into polynomial.rs
